Use double (instead float) + performance (#8)

+++ include/ifcpp/Geometry/Matrix.h
@@ -12,17 +12,17 @@ template<CVector TVector>
 class Matrix {
     using AVector = VectorAdapter<TVector>;
 public:
-    float data[ 4 ][ 4 ];
+    double data[ 4 ][ 4 ];
     static Matrix<TVector> GetIdentity() {
         Matrix<TVector> matrix {};
-        memset( matrix.data, 0, sizeof( float ) * 16 );
+        memset( matrix.data, 0, sizeof( double ) * 16 );
         matrix.data[ 0 ][ 0 ] = 1;
         matrix.data[ 1 ][ 1 ] = 1;
         matrix.data[ 2 ][ 2 ] = 1;
         matrix.data[ 3 ][ 3 ] = 1;
         return matrix;
     }
-    static Matrix<TVector> GetScale( float x, float y, float z ) {
+    static Matrix<TVector> GetScale( double x, double y, double z ) {
         auto matrix = Matrix::GetIdentity();
         matrix.data[ 0 ][ 0 ] = x;
         matrix.data[ 1 ][ 1 ] = y;
@@ -54,33 +54,33 @@ public:
 
         return matrix;
     }
-    static Matrix<TVector> GetRotation( float angle, const TVector& v ) {
+    static Matrix<TVector> GetRotation( double angle, const TVector& v ) {
         auto result = Matrix::GetIdentity();
 
         auto a = angle;
-        auto c = cosf( a );
-        auto s = sinf( a );
+        auto c = cos( a );
+        auto s = sin( a );
         auto axis = AVector::Normalized( v );
 
-        result.data[ 0 ][ 0 ] = c + ( 1.0f - c ) * axis.x * axis.x;
-        result.data[ 0 ][ 1 ] = ( 1.0f - c ) * axis.x * axis.y + s * axis.z;
-        result.data[ 0 ][ 2 ] = ( 1.0f - c ) * axis.x * axis.z - s * axis.y;
-        result.data[ 0 ][ 3 ] = 0.0f;
+        result.data[ 0 ][ 0 ] = c + ( 1.0 - c ) * axis.x * axis.x;
+        result.data[ 0 ][ 1 ] = ( 1.0 - c ) * axis.x * axis.y + s * axis.z;
+        result.data[ 0 ][ 2 ] = ( 1.0 - c ) * axis.x * axis.z - s * axis.y;
+        result.data[ 0 ][ 3 ] = 0.0;
 
-        result.data[ 1 ][ 0 ] = ( 1.0f - c ) * axis.y * axis.x - s * axis.z;
-        result.data[ 1 ][ 1 ] = c + ( 1.0f - c ) * axis.y * axis.y;
-        result.data[ 1 ][ 2 ] = ( 1.0f - c ) * axis.y * axis.z + s * axis.x;
-        result.data[ 1 ][ 3 ] = 0.0f;
+        result.data[ 1 ][ 0 ] = ( 1.0 - c ) * axis.y * axis.x - s * axis.z;
+        result.data[ 1 ][ 1 ] = c + ( 1.0 - c ) * axis.y * axis.y;
+        result.data[ 1 ][ 2 ] = ( 1.0 - c ) * axis.y * axis.z + s * axis.x;
+        result.data[ 1 ][ 3 ] = 0.0;
 
-        result.data[ 2 ][ 0 ] = ( 1.0f - c ) * axis.z * axis.x + s * axis.y;
-        result.data[ 2 ][ 1 ] = ( 1.0f - c ) * axis.z * axis.y - s * axis.x;
-        result.data[ 2 ][ 2 ] = c + ( 1.0f - c ) * axis.z * axis.z;
-        result.data[ 2 ][ 3 ] = 0.0f;
+        result.data[ 2 ][ 0 ] = ( 1.0 - c ) * axis.z * axis.x + s * axis.y;
+        result.data[ 2 ][ 1 ] = ( 1.0 - c ) * axis.z * axis.y - s * axis.x;
+        result.data[ 2 ][ 2 ] = c + ( 1.0 - c ) * axis.z * axis.z;
+        result.data[ 2 ][ 3 ] = 0.0;
 
         return result;
     }
     void Transform( TVector* v ) const {
-        float rw = 1.0f / ( this->data[ 3 ][ 0 ] * v->x + this->data[ 3 ][ 1 ] * v->y + this->data[ 3 ][ 2 ] * v->z + this->data[ 3 ][ 3 ] );
+        double rw = 1.0 / ( this->data[ 3 ][ 0 ] * v->x + this->data[ 3 ][ 1 ] * v->y + this->data[ 3 ][ 2 ] * v->z + this->data[ 3 ][ 3 ] );
         *v = { ( this->data[ 0 ][ 0 ] * v->x + this->data[ 0 ][ 1 ] * v->y + this->data[ 0 ][ 2 ] * v->z + this->data[ 0 ][ 3 ] ) * rw,
                ( this->data[ 1 ][ 0 ] * v->x + this->data[ 1 ][ 1 ] * v->y + this->data[ 1 ][ 2 ] * v->z + this->data[ 1 ][ 3 ] ) * rw,
                ( this->data[ 2 ][ 0 ] * v->x + this->data[ 2 ][ 1 ] * v->y + this->data[ 2 ][ 2 ] * v->z + this->data[ 2 ][ 3 ] ) * rw };
